+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Architectural Pattern Helpers
 Extensions
 
 - `Optional` `String` method `compactMapNonEmpty` is renamed to `compactMapNonNilNonEmpty`
+- `Calendar` can now calculate number of days in month from year and month as arguments
 
 #### [3.3.1(24)](https://github.com/VakhoKontridze/VCore/releases/tag/3.3.1) — *2022 06 08*